move the functions in their respective scopes which are defined as files
